+++ src/rbm.hpp
@@ -33,24 +33,24 @@ class RBM : public Model {
         void resetVisProbs();
         void sumVisProbs();
         void sumToVisProbs();
-        double getActualVal(int n, int i, int j, int k);
-        double getExpectVal(int n, int i, int j, int k);
+        float getActualVal(int n, int i, int j, int k);
+        float getExpectVal(int n, int i, int j, int k);
         void runGibbsSampler();
         void train(std::string saveFile);
 
     private:
-        double* W;
-        double* dW; // So we don't have to reallocate each time.
+        float* W;
+        float* dW; // So we don't have to reallocate each time.
         unsigned int T; // See equation 6 in RBM for CF, Salakhutdinov 2007
-        double epsilonW; // learning rate for weights
-        double epsilonVB; // learning rate for biases of visible units
-        double epsilonHB; // learning rate for biases of hidden units
-        double* hidBiases; // bias of feature j
-        double* dHidBiases; // delta of bias of feature j
-        double* visBiases; // bias of rating k for movie i
-        double* dVisBiases; // delta of bias of rating k for movie i
-        double* hidProbs; // hidden probabilities
-        double* visProbs; // visible probabilities
+        float epsilonW; // learning rate for weights
+        float epsilonVB; // learning rate for biases of visible units
+        float epsilonHB; // learning rate for biases of hidden units
+        float* hidBiases; // bias of feature j
+        float* dHidBiases; // delta of bias of feature j
+        float* visBiases; // bias of rating k for movie i
+        float* dVisBiases; // delta of bias of rating k for movie i
+        float* hidProbs; // hidden probabilities
+        float* visProbs; // visible probabilities
         std::bitset<N_USERS * N_FACTORS>* hidVars; // bitset
         std::bitset<N_MOVIES * MAX_RATING>* indicatorV; // array of bitsets
 };
